% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultra_sparse_backends.R
\name{bvs_mh_singlenet_ultra_sparse}
\alias{bvs_mh_singlenet_ultra_sparse}
\title{Ultra-sparse MH BVS with single sparse GGM adjacency}
\usage{
bvs_mh_singlenet_ultra_sparse(
  X,
  y,
  S_ggm = NULL,
  niter = 60000L,
  burnin = 10000L,
  thin = 1L,
  nu0 = 2,
  sigmasq0 = 1.5,
  h = 1.5,
  mu = -log(1/0.3 - 1),
  alpha0 = 0,
  beta0 = 0,
  n_mh_gamma = 3L,
  eta_sd = 0.5,
  mu_tilde = -4,
  eta1_tilde = 0.075,
  e_eta = 2,
  f_eta = 1,
  Tmax = 64L,
  proposal_type = 1L,
  v0_ggm = 0.015^2,
  v1_ggm = (50^2) * (0.015^2),
  pii_ggm = NULL,
  beta_init = NULL,
  gamma_init = NULL,
  alpha_init = NULL,
  store_beta = FALSE,
  store_gamma = FALSE,
  store_Z_list = FALSE,
  store_Z_pip = TRUE,
  seed = NULL
)
}
\description{
Ultra-sparse MH BVS with single sparse GGM adjacency
}
