% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_summary.R
\name{.extract_pip_from_bvs}
\alias{.extract_pip_from_bvs}
\title{Summarize BVS Results}
\usage{
\method{summary}{bvs}(object)
}
\arguments{
\item{object}{An object of class \code{"bvs"}.}

\item{pip_threshold}{Threshold on PIP for declaring a variable selected
(default 0.5).}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with:
  \describe{
    \item{pip}{Posterior inclusion probabilities (length p).}
    \item{selected}{Indices of selected variables (PIP > threshold).}
    \item{beta_mean}{Posterior mean of beta (length p).}
    \item{beta_lower}{2.5\% quantile of beta (length p).}
    \item{beta_upper}{97.5\% quantile of beta (length p).}
    \item{alpha_mean}{Posterior mean of alpha.}
    \item{nselected}{Number of selected variables.}
  }
}
\description{
Computes posterior inclusion probabilities (PIP), posterior means,
and 95\% credible intervals for each predictor.
}
\examples{
\dontrun{
fit <- bvs_pg(X, y, adj_type = "fixed", adj_fixed = R, niter = 5000)
s <- summary(fit)
s$selected
}

}
