% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.init_mcmc}
\alias{.init_mcmc}
\title{Initialize MCMC Parameters for BVS}
\usage{
.init_mcmc(p, mu, nu0, sigmasq0, alpha0, beta0, h)
}
\arguments{
\item{p}{Number of predictors.}

\item{mu}{Ising external field.}

\item{nu0}{IG shape parameter.}

\item{sigmasq0}{IG scale parameter.}

\item{alpha0}{Prior mean for intercept.}

\item{beta0}{Prior mean for coefficients.}

\item{h}{Intercept variance inflation.}
}
\value{
A named list with components: beta_init, gamma_init, alpha_init.
}
\description{
Generates default initial values for beta, gamma, alpha, and sigma^2.
}
\keyword{internal}
