% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_summary.R
\name{print.bvs}
\alias{print.bvs}
\title{Print BVS Object}
\usage{
\method{print}{bvs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bvs"}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print BVS Object
}
