% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BVS.DAdj-package}
\alias{BVS.DAdj}
\alias{BVS.DAdj-package}
\title{BVS.DAdj: Bayesian Variable Selection with Dual-Adjacency Network-Informed Ising Priors}
\description{
MCMC methods for Bayesian variable selection in logistic regression with Ising/MRF priors informed by graphical model adjacency. Supports Metropolis-Hastings and Polya-Gamma augmented Gibbs sampling with fixed, graphical lasso, and Bayesian GGM adjacency structures. Includes dense and sparse high-dimensional variants with single and dual network priors.

MCMC tools for Bayesian variable selection in logistic
  regression with Ising/MRF priors.
}
\author{
\strong{Maintainer}: Yubing Yao \email{yyao@umass.edu}

Authors:
\itemize{
  \item Mahlet Tadesse \email{mgt26@georgetown.edu}
  \item Raji Balasubramanian \email{rbalasub@schoolph.umass.edu}
}

}
\keyword{internal}
