% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_summary.R
\name{plot.bvs}
\alias{plot.bvs}
\title{Plot BVS Posterior Inclusion Probabilities}
\usage{
\method{plot}{bvs}(x, pip_threshold = 0.5, top_n = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"bvs"}.}

\item{pip_threshold}{Threshold line (default 0.5).}

\item{top_n}{Show only top N variables by PIP (default: all).}

\item{...}{Additional graphical arguments.}
}
\description{
Creates a barplot of posterior inclusion probabilities with a threshold line.
}
\examples{
\dontrun{
fit <- bvs_pg(X, y, adj_type = "fixed", adj_fixed = R, niter = 5000)
plot(fit)
}

}
